added trait
