+++ stats/stats/src/lib.rs
@@ -2,4 +2,4 @@ mod read;
 mod update;
 
 pub use read::{get_chart_int, get_counters, ReadError};
-pub use update::update_new_blocks_per_day;
+pub use update::{new_blocks, UpdaterTrait};
